pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.116.1",
    "langchain-google-genai>=2.1.9",
    "langchain-openai>=0.3.30",
    "langgraph>=0.6.5",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.35.0",
]
